contract input & output format with decimals, resolves #291
