+++ src/app/modules/wallet/wallet-send/wallet-send.component.html
@@ -148,7 +148,15 @@
                      *ngFor="let functionInput of selectedFunction.inputs; let i=index"
                      class="uk-margin">
                   <div class="uk-form-controls">
-                    <input class="uk-input" placeholder="{{functionInput.name}}" formControlName="{{i}}">
+                    <ng-container *ngIf="selectedFunction.name === 'transfer' && functionInput.name === 'value' && addr.token_symbol; else regularFunctionInput">
+                      <input type="number" class="uk-input" placeholder="Value ({{addr.token_symbol}})" formControlName="{{i}}" #valueInput>
+                      <p *ngIf="valueInput.value">
+                        Raw value: {{valueInput.value * ('1e' + addr.decimals)}}
+                      </p>
+                    </ng-container>
+                    <ng-template #regularFunctionInput>
+                      <input class="uk-input" placeholder="{{functionInput.name}}" formControlName="{{i}}">
+                    </ng-template>
                   </div>
                 </div>
                 <div *ngIf="selectedFunction.payable" class="uk-margin">
