+++ src/app/pipes/big-number.pipe.ts
@@ -1,4 +1,5 @@
 import {PipeTransform, Pipe} from '@angular/core';
+import {numberWithCommas} from '../utils/functions';
 
 @Pipe({
   name: 'bigNumber'
@@ -7,11 +8,6 @@ import {PipeTransform, Pipe} from '@angular/core';
 export class BigNumberPipe implements PipeTransform {
 
   transform(val: string): string {
-    if (val == null) {
-      return val;
-    }
-    const parts = val.toString().split('.');
-    parts[0] = parts[0].replace(/\B(?=(\d{3})+(?!\d))/g, ',');
-    return parts.join('.');
+    return numberWithCommas(val);
   }
 }
